+++ src/main/java/roomescape/dto/ReservationResponseDTO.java
@@ -1,0 +1,18 @@
+package roomescape.dto;
+
+import java.time.LocalDate;
+import roomescape.domain.Reservation;
+
+public record ReservationResponseDTO(
+        String name,
+        LocalDate date,
+        Long timeId
+) {
+    public static ReservationResponseDTO from(Reservation reservation) {
+        return new ReservationResponseDTO(
+                reservation.getName(),
+                reservation.getDate(),
+                reservation.getTime().getId()
+        );
+    }
+}
