+++ src/main/java/roomescape/dto/ReservationTimeRequestDTO.java
@@ -1,0 +1,7 @@
+package roomescape.dto;
+
+public record ReservationTimeRequestDTO(
+        String startAt
+) {
+
+}
